+++ codifyiq_core_components/lib/codifyiq_core_components.dart
@@ -14,3 +14,4 @@ export 'widgets/pdf_viewer_widget.dart';
 export 'widgets/image_viewer_widget.dart';
 export 'widgets/social_sign_in_screen.dart';
 export 'widgets/terms_and_conditions_widget.dart';
+export 'widgets/user_avatar.dart';